extract apply logic to support class
